+++ gto/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # coding=utf8
 
-from .tobe import ISSH, CONFIG_PATH
+from .gto import ISSH, CONFIG_PATH
 
 def main():
     ssh = ISSH(CONFIG_PATH)
