+++ gto/version.py
@@ -1,0 +1,3 @@
+__version_info__ = (0, 1, 0)
+__version__ = ".".join(map(str, __version_info__))
+__createdate__ = '2020-03-20'
