--- tobe/version.py
@@ -1,3 +1,0 @@
-__version_info__ = (0, 4, 2)
-__version__ = ".".join(map(str, __version_info__))
-__createdate__ = '2020-02-09'
